centralize util for making an empty directory
